web-panel/backend/requirements.txt
# ===================================================================
# VPSService Web Panel - Backend Requirements
# Django REST Framework para API REST
# Autenticación JWT para seguridad
# CORS para conexiones desde frontend
# ===================================================================

Django==4.2.13
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.2
python-decouple==3.8
psycopg2-binary==2.9.9
Pillow==10.1.0
celery==5.3.4
redis==5.0.1
paramiko==3.4.0
fabric==3.2.2

# Utilidades de desarrollo (opcional)
django-debug-toolbar==4.2.0
black==23.12.1
flake8==6.1.0
pytest==7.4.3
pytest-django==4.7.0
